Modify all Makefiles to reference `build_rule.mk`.

+++ current/src/urg_tcpclient.c
@@ -61,7 +61,7 @@ static int32_t tcpclient_buffer_read(urg_tcpclient_t *cli, char *data, int32_t s
 static void set_block_mode(urg_tcpclient_t *cli)
 {
 #if defined(URG_WINDOWS_OS)
-    uint32_t flag = 0;
+    u_long flag = 0;
     ioctlsocket(cli->sock_desc, FIONBIO, &flag);
 #else
     int32_t flag = 0;
@@ -78,7 +78,7 @@ int32_t tcpclient_open(urg_tcpclient_t *cli, const char *ip_str, int32_t port_nu
     fd_set rmask, wmask;
     struct timeval tv = {Connect_timeout_second, 0};
 #if defined(URG_WINDOWS_OS)
-    uint32_t flag;
+    u_long flag;
 #else
     int32_t flag;
     int32_t sock_optval = -1;
@@ -260,7 +260,7 @@ int32_t tcpclient_read(urg_tcpclient_t *cli,
         char tmpbuf[BUFSIZE];
         // receive with non-blocking mode.
 #if defined(URG_WINDOWS_OS)
-        uint32_t val = 1;
+        u_long val = 1;
         ioctlsocket(sock, FIONBIO, &val);
         n = recv(sock, tmpbuf, BUFSIZE - num_in_buf, 0);
 #else
@@ -283,7 +283,7 @@ int32_t tcpclient_read(urg_tcpclient_t *cli,
     //  lastly recv with blocking but with time out to read necessary size.
     {
 #if defined(URG_WINDOWS_OS)
-        uint32_t val = 0;
+        u_long val = 0;
         ioctlsocket(sock, FIONBIO, &val);
         setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO,
                    (const char *)&timeout, sizeof(struct timeval));
